requirements.txt
google-genai>=0.7.0
pydantic>=2.6.0
python-dotenv>=1.0.1
pandas>=2.2.0
numpy>=1.26.0
tqdm>=4.66.0